Change: move create_user to dedicated users files
