+++ src/manage.h
@@ -3215,11 +3215,6 @@ gchar *
 keyfile_to_auth_conf_settings_xml (const gchar *);
 
 int
-create_user (const gchar *, const gchar *, const gchar *, const gchar *,
-             int, const array_t *, array_t *, gchar **,
-             array_t *, gchar **, gchar **, user_t *, int);
-
-int
 delete_user (const char *, const char *, int, const char*, const char*);
 
 int
